Add test containers
This commit adds two more containers to the dev setup:

- test
- db_test

This way tests can be executed in a completely separate environment
from the development containers which makes it more felxible.

It adds one extra step to the setup:

``` bash
$ doco run test make setup
```

Running tests:

``` bash
$ doco run test
```

Downside is that it starts an extra MySQL container.

Also, in many cases MySQL might not even be necessary for running tests
(eg. Liip FunctionalTestBundle is used with SQLite), but that's
probably not the majority and for integration tests MySQL is
necessary anway.